Added w32toggletitle.exe to the repo. Used for fullscreen.
Changed the path to includes using ~.
The path to the exe had to be updated in w32-fullscreen.el.

Relative path was not working here, since the current path to
init.el is not the same as that of emacs.

Future possibility: set the path to includes and
w32toggletitle.exe relatively to init.el, instead of ~.